+++ assets/foliate-js/book.js
@@ -165,6 +165,9 @@ const setSelectionHandler = (view, doc, index) => {
   ) {
     doc.addEventListener('pointerup', () => handleSelection(view, doc, index));
   }
+  else {
+    doc.addEventListener('contextmenu', () => handleSelection(view, doc, index));
+  }
   // doc.addEventListener('selectionchange', () => handleSelection(view, doc, index));
 
   if (!view.isFixedLayout) {
@@ -778,7 +781,10 @@ const open = async (file, cfi) => {
 }
 
 
-const callFlutter = (name, data) => window.flutter_inappwebview.callHandler(name, data)
+const callFlutter = (name, data) => {
+  // console.log('callFlutter', name, data)
+  window.flutter_inappwebview.callHandler(name, data)
+}
 
 const setStyle = () => {
   const turn = {