+++ lib/config/shared_preference_provider.dart
@@ -557,4 +557,28 @@ class Prefs extends ChangeNotifier {
     prefs.setBool('showTextUnderIconButton', show);
     notifyListeners();
   }
+
+  DateTime? get lastUploadBookDate {
+    String? lastUploadBookDateStr = prefs.getString('lastUploadBookDate');
+    if (lastUploadBookDateStr == null) return null;
+    return DateTime.parse(lastUploadBookDateStr);
+  }
+
+  set lastUploadBookDate(DateTime? date) {
+    if (date == null) {
+      prefs.remove('lastUploadBookDate');
+    } else {
+      prefs.setString('lastUploadBookDate', date.toIso8601String());
+    }
+    notifyListeners();
+  }
+
+  int get lastServerPort {
+    return prefs.getInt('lastServerPort') ?? 0;
+  }
+
+  set lastServerPort(int port) {
+    prefs.setInt('lastServerPort', port);
+    notifyListeners();
+  }
 }