+++ lib/page/settings_page/more_settings_page.dart
@@ -5,6 +5,7 @@ import 'package:anx_reader/page/settings_page/appearance.dart';
 import 'package:anx_reader/page/settings_page/narrate.dart';
 import 'package:anx_reader/page/settings_page/reading.dart';
 import 'package:anx_reader/page/settings_page/settings_page.dart';
+import 'package:anx_reader/page/settings_page/storege.dart';
 import 'package:anx_reader/page/settings_page/sync.dart';
 import 'package:anx_reader/page/settings_page/translate.dart';
 import 'package:anx_reader/widgets/settings/about.dart';
@@ -79,6 +80,7 @@ class _SubMoreSettingsState extends State<SubMoreSettings> {
             "sections": const ReadingSettings(),
             "subtitles": [
               L10n.of(context).reading_page_reading,
+              L10n.of(context).download_fonts,
               L10n.of(context).reading_page_style,
               L10n.of(context).reading_page_other,
             ],
@@ -116,6 +118,15 @@ class _SubMoreSettingsState extends State<SubMoreSettings> {
             "subtitles": [
               L10n.of(context).settings_ai_services,
               L10n.of(context).settings_ai_prompt,
+            ],
+          },
+          {
+            "title": L10n.of(context).storage,
+            "icon": Icons.storage_outlined,
+            "sections": const StorageSettings(),
+            "subtitles": [
+              L10n.of(context).storage_info,
+              L10n.of(context).storage_data_file_details,
             ],
           },
           {